requirements.txt
antlr4-python3-runtime==4.9.3
anyio==4.11.0
audioread==3.1.0
bitarray==3.8.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
Cython==3.2.0
decorator==5.2.1
exceptiongroup==1.3.0
fairseq @ https://github.com/One-sixth/fairseq/archive/main.zip#sha256=6f876d6c550b49dcc74ac0756806efe36aeb17458ccc33b4d972291a0617061c
faiss-cpu==1.12.0
filelock==3.19.1
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.0.1
hydra-core==1.3.2
idna==3.11
inferrvc @ https://github.com/CircuitCM/RVC-inference/raw/main/dist/inferrvc-1.0-py3-none-any.whl#sha256=12d1c074dc4578956c33f3f59dad708929ebdbf8a4e29b15144c09059b1f6ea0
Jinja2==3.1.6
joblib==1.5.2
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.45.1
lxml==6.0.2
MarkupSafe==2.1.5
mpmath==1.3.0
msgpack==1.1.2
networkx==3.3
numba==0.62.1
numpy==2.1.2
omegaconf==2.3.0
packaging==25.0
pillow==11.3.0
platformdirs==4.5.0
pooch==1.8.2
portalocker==3.2.0
praat-parselmouth==0.4.6
pycparser==2.23
pywin32==311
pyworld==0.3.5
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
resampy==0.4.3
sacrebleu==2.5.1
scikit-learn==1.7.2
scipy==1.15.3
shellingham==1.5.4
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
torch==2.9.0+cpu
torchaudio==2.9.0+cpu
torchcrepe==0.0.24
torchvision==0.24.0+cpu
tqdm==4.67.1
typer-slim==0.20.0
typing_extensions==4.15.0
urllib3==2.5.0
